Refactor job/task system
This is a major refactoring of the job and task system. Jobs can now
be thought of as a container for an arbitrary list of tasks. The new
system supports a more flexible approach to defining different job
types. Tasks are self-contained functions that can be reused.

Additionally, rather than accepting an arbitrary number of files of
any type, the application expects a single Bag upon submission.