Add import checker
This adds a task to poll GeoServer for the status of an import task and
updates the job if the import has been completed. The `import_check.py`
script will queue this task to be run by the RQ worker. It's intended to
be run regularly on a cron or similar.